2017-08-02  company commit

+++ app/src/main/java/com/thisrpg/SQLdm.java
@@ -5,8 +5,17 @@ package com.thisrpg;
  */
 
 import android.content.Context;
+import android.content.pm.PackageInfo;
+import android.content.pm.PackageManager;
 import android.content.res.AssetManager;
+import android.database.DatabaseErrorHandler;
 import android.database.sqlite.SQLiteDatabase;
+import android.database.sqlite.SQLiteOpenHelper;
+
+import com.thisrpg.db.DaoMaster;
+import com.thisrpg.db.DaoSession;
+
+import org.greenrobot.greendao.database.Database;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -15,18 +24,35 @@ import java.io.InputStream;
 
 /**
  * 这个类就是实现从assets目录读取数据库文件然后写入SDcard中,如果在SDcard中存在，就打开数据库，不存在就从assets目录下复制过去
- *
- * @author Big_Adamapple
  */
 public class SQLdm {
 
-    //数据库存储路径
-    private static final String filePath = "data/data/com.thisrpg/rpg.db";
-    //数据库存放的文件夹 data/data/com.main.jh 下面
-    private static final String pathStr = "data/data/com.thisrpg";
 
-    interface LoadDataCallback {
-        void loadSuccess(SQLiteDatabase database);
+    private static DaoMaster daoMaster;
+    private static DaoSession daoSession;
+
+    public static DaoMaster getDaoMaster() {
+        return daoMaster;
+    }
+
+    public static void setDaoMaster(DaoMaster daoMaster) {
+        SQLdm.daoMaster = daoMaster;
+    }
+
+    public static DaoSession getDaoSession() {
+        return daoSession;
+    }
+
+    public static void setDaoSession(DaoSession daoSession) {
+        SQLdm.daoSession = daoSession;
+    }
+
+    //数据库存放的文件夹 data/data/com.main.jh 下面
+    private static final String pathStr = "data/data/com.thisrpg/databases";
+
+
+    public interface LoadDataCallback {
+        void loadSuccess();
 
         //-1：文件夹创建失败
         //-2：assets中的数据库文件导出失败
@@ -36,30 +62,42 @@ public class SQLdm {
 
     public static void loadData(final Context context, final String databaseName, final LoadDataCallback l) {
 
-
         File path = new File(pathStr);
         LogUtils.e("pathStr=" + path);
         //不存在先创建文件夹
-        if (!path.mkdir()) {
-            l.loadFail(-1);
+        if (!path.exists()) {
+            if (!path.mkdir()) {
+                l.loadFail(-1);
+                return;
+            }
         }
 
         new Thread(new Runnable() {
             @Override
             public void run() {
-                //循环需要导出的文件名
+
+                //判断是否需要更新数据库
                 File file = new File(pathStr + "/" + databaseName);
-                if (!SharedPreferencesUtils.getString(context, "version", "").equals(SystemUtils.getVersion(context))) {
-                    if (!file.delete()) {
-                        l.loadFail(-3);
+                //判断文件如果存在
+                if (file.exists()) {
+                    //那么检测版本号是否一致
+                    if (!SharedPreferencesUtils.getString(context, "version", "").equals(SystemUtils.getVersion(context))) {
+                        //版本号不一致则需要更新，将上一个版本的数据库删除之后再重新初始化
+                        if (!file.delete()) {
+                            l.loadFail(-3);
+                            return;
+                        }
                     }
                 }
+
 
                 //之后删掉这个判断
                 if (file.exists()) {
                     file.delete();
                 }
 
+                //检测完版本号之后再判定数据库文件是否存在
+                //如果不存在，则重新从assets文件夹中将数据库文件导出本地
                 if (!file.exists()) {
                     InputStream is = null;
                     FileOutputStream fos = null;
@@ -100,9 +138,28 @@ public class SQLdm {
                         }
                     }
                 }
+                //将版本号保存到本地
                 SharedPreferencesUtils.editorPut(context, "version", SystemUtils.getVersion(context));
-                l.loadSuccess(SQLiteDatabase.openOrCreateDatabase(pathStr + "/" + databaseName, null));
+                //实例化DaoMaster，DaoSession
+                if (daoMaster == null) {
+                    daoMaster = new DaoMaster(new DaoMaster.DevOpenHelper(context, databaseName).getReadableDatabase());
+                    daoSession = daoMaster.newSession();
+                }
+                l.loadSuccess();
             }
         }).start();
+    }
+
+    private PackageInfo getAppInfo(Context context) {
+        PackageInfo packageInfo = null;
+        // 获取packageManager的实例
+        PackageManager packageManager = context.getPackageManager();
+        // getPackageName()是你当前类的包名，0代表是获取版本信息
+        try {
+            packageInfo = packageManager.getPackageInfo(context.getPackageName(), 0);
+        } catch (PackageManager.NameNotFoundException e) {
+            e.printStackTrace();
+        }
+        return packageInfo;
     }
 }
